bhyve: Move MSR emulation into amd64/
No functional change intended.

Reviewed by:	corvink, jhb
MFC after:	1 week
Sponsored by:	Innovate UK
Differential Revision:	https://reviews.freebsd.org/D40733

(cherry picked from commit 548b11228f7a83785dcf6c4c8b0bbfed0284d794)
