+++ usr.sbin/bhyve/amd64/Makefile.inc
@@ -7,7 +7,8 @@ SRCS+=	\
 	post.c		\
 	ps2kbd.c	\
 	ps2mouse.c	\
-	task_switch.c
+	task_switch.c	\
+	xmsr.c
 
 .PATH:  ${BHYVE_SYSDIR}/sys/amd64/vmm
 SRCS+=	vmm_instruction_emul.c
