+++ usr.sbin/bhyve/amd64/xmsr.c
